% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook.syn.r
\name{codebook.syn}
\alias{codebook.syn}
\title{Makes a codebook from a data frame}
\usage{
codebook.syn(data, maxlevs = 3)
}
\arguments{
\item{data}{a data frame with a data set to be synthesised.}

\item{maxlevs}{the number of factor levels above which separate tables with
all labels are returned as part of \code{labs} component.}
}
\value{
a list with two components:
\itemize{
\item \strong{tab}: a data frame with the following information about each variable:
\itemize{
\item \strong{name}: variable name
\item \strong{class}: class of variable
\item \strong{nmiss}: number of missing values (\code{NA})
\item \strong{perctmiss}: percentage of missing values
\item \strong{ndistinct}: number of distinct values (excluding missing values)
\item \strong{details}: range for numeric variables, maximum length for character
variables, labels for factors with <= maxlevs levels
}
\item \strong{labs}: a list of extra tables with labels for each factor with a
number of levels greater than \code{maxlevs}.
}
}
\description{
Describes features of variables in a data frame relevant for synthesis.
}
\examples{
codebook.syn(SD2011)
}
