% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.syn.r
\name{compare}
\alias{compare}
\title{Comparison of synthesised and observed data}
\usage{
compare(object, data, ...)
}
\arguments{
\item{object}{a synthetic data object of class \code{synds}
or \code{fit.synds}.}

\item{data}{an original observed data set.}

\item{...}{additional arguments specific to a method.}
}
\value{
The value returned by \code{compare} depends on the class of its argument.
See the documentation of the particular methods for details.
}
\description{
A generic function for the comparison of synthesised and observed data.
The function invokes particular methods that depend on the class of the
first argument.
}
\details{
Compare methods facilitate quality assessment of synthetic data by comparing
them with the original observed data sets. The data themselves (for class
\code{synds}) or models fitted to them (for class \code{fit.synds}) are
compared.
}
\examples{
\dontrun{
compare(synthetic_data, observed_data)
}

}
\seealso{
\code{compare.synds}, \code{compare.fit.synds}
}
